Cambio de tipo de dato al atributo totalRecords en la clase TableDTO

+++ src/main/java/com/cltbackend/dto/TableDTO.java
@@ -11,7 +11,7 @@ public class TableDTO<T> {
 
     private List<T> lista;
 
-    private int totalRecords;
+    private long totalRecords;
 
    public TableDTO(){
